requirements-old.txt
anonymizeip==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cx-Oracle==8.1.0
decorator==4.4.2
doltpy==1.1.10
future==0.18.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
IP2Location==8.5.1
IP2Proxy==3.0.0
joblib==1.0.0
mysql-connector-python==8.0.21
mysqlclient==2.0.3
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.1
Pillow==8.1.0
priority==1.3.0
protobuf==3.14.0
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.10.0
PyPika==0.47.4
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retry==0.9.2
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.14.1
subresource-integrity==0.2
textblob==0.15.3
toml==0.10.2
tqdm==4.56.0
urllib3==1.26.2
wsproto==1.0.0
youtube-dl==2021.1.24.1
